requirements.txt
aiogram==3.13.1
pandas==2.2.2
openpyxl==3.1.2
apscheduler==3.10.4
python-dotenv==1.0.1